agregando policies, politicas para verificacion de usuario/post al momento de acceder a contenido permitido o no y restriccion por pertenencia al contenido de cada usuario

+++ app/Http/Controllers/Admin/PostController.php
@@ -90,6 +90,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {   
+        
+        $this->authorize('author', $post);
+
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
 
@@ -104,7 +107,9 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(PostRequest $request, Post $post)
-    {
+    {   
+        $this->authorize('author', $post);
+
         $post->update($request->all());
 
         if($request->file('file')){
@@ -140,7 +145,9 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Post $post)
-    {
+    {   
+        $this->authorize('author', $post);
+        
         $post->delete();
         return redirect()->route('admin.posts.index')->with('info', 'El post se elimino con exito!');
     }
